requirements.txt
streamlit
python-docx
spacy
sv-core-news-sm